Remove coupon proto-generated files.
Deleted proto-generated Go files related to the coupon system, including client, server, and message definitions. This cleanup ensures no stale or unnecessary files remain in the repository.

+++ common/id/id.go
@@ -4,7 +4,7 @@ import "sync"
 
 type ID struct {
 	mu sync.Mutex
-	n  uint64
+	n  uint32
 }
 
 // NewID creates and returns a pointer to a new ID instance initialized with zero.
@@ -12,8 +12,8 @@ func NewID() *ID {
 	return &ID{n: 0}
 }
 
-// Next generates and returns the next sequential uint64 value in a thread-safe manner.
-func (i *ID) Next() uint64 {
+// Next generates and returns the next sequential uint32 value in a thread-safe manner.
+func (i *ID) Next() uint32 {
 	i.mu.Lock()
 	defer i.mu.Unlock()
 	i.n++
